ContactList Commit 4
This Commit provides functionality that allows the user to enter input
into the input fields, when the user adds this information and clicks
the button an instance of the contact class is created accepting input
fields as the parameters, before a new contact is created all of the
fields are checked for blank input. Once a new contact has been created
a list view item is created and its values are set to the values of the
contact.
Text boxes are cleared and blank input message is shown if there are
blank fields.