ContactList Commit 6
This Commit provides functionality that checks each text box. There are
4 new methods, a leave method for each textbook. When the textboxes are
selected then unselected the CheckInput method is called with that
textboxes text value as the parameter. If the method returns true than
the program continues. If the method return is false then a message box
pops up to show invalid input and the textbox text is highlighted for
revision.