ContactList Commit 7
This commit provides functionality that allows the user to change the
size of the icons for the list view items. When large icon is selected
the listView items icon size is set to large and the button is checked.
When the small icon is selected the listView items icon size is set to
small and the button is checked.