ContactList Commit 11
This commit provides functionality that allows the user to save the
contactList. When the user clicks the save button a save file dialog is
opened and the user can create a file name and select the file
location. All of the list view items are saved to a text file.