TicTacToe Commit 1
This commit Provides the basics of functionality for this program. Some
variables have been created to store data and keep track of game
variables and 2 methods have been created. The method for when the
player selects X and the method for when the player selects O. when the
player selects X a method sets a global variable called imagined to 1
and if O is selected then the imagined variable is set to 0, this is
for mathematical purposes and will allow us to keep track of whose turn
it is.