TicTacToe Commit 2
This commit does allot for the program. There are 9 new methods but
they all provide the same function just for different tiles on the
board. The first part of each method checks that either X or O has been
selected, if the variable imagined is -1 then X  or O have not been
selected. Then it will check if any player has already played this tile
by checking a corresponding boolean in a boolean array. If both
conditions are met then that tile will be marked true in the boolean
array, The image index is then set using an algorithm that will get the
correct X or O image index by adding the number of plays counter to the
imgIndex variable then modulas 2. The answer is set to the buttons
image index and the calculated image index is saved in a 2d array for
scoring, the counter incremented.