TicTacToe Commit 3
This commit provides functionality that allows the user to change the
game pieces from blue to red or red to blue. There are 2 methods added
to the program to do this. Either method simply changes all of the
buttons image lists to either the blue image list or red image list
based on which was selected, then checks blue or red accordingly.