TicTacToe Commit 4
This commit provides functionality that allows the user to start a new
game. To do this 2 new methods are added, one called newGame and one to
handle the new game button. The new game button simply calls the new
game method and this is because there is more than one instance where a
new game needs to be created. The new game method sets all imageLists
back to blue which is the default, sets the variable imagIndex back to
-1, resets the counter, the boolean array, and the 2d array.