TicTacToe Commit 5
This Commit contains 1 additional method that checks the score of the
game. This method loops through the 2d array that holds player moves
and compares the location of each player tile to look for 3 in a row.
If there are 3 of a players tile in a row then the players imgIndex
variable is returned. If the counter is at max and there is not winner
then a variable is returned to represent a tie.The method is called
every time a player selects a tile.