TicTacToe Commit 6
This commit has 1 additional method to take in the results of the
checkscore method and provide output that displays the winner. So the
checkscore method is called within the showWinner method, 1 of 3 things
will be returned from the check score method, either a -1 for no winner
yet, either two of the player imgIndex variables if either gets three
in a row. The show winner method checks the returned variables to
display who won or if it was a tie, If someone won or it was a tie then
the result is displayed and a new game is created.