GroceryList3 Commit 4
This commit provides functionality for the switch button. There is a
new method for the switch button click. It takes the selected item from
the focused list view and removes it from both the list and list view.
It then adds that item to the other list and list view.