ContactList3 Commit 4
This commit provides functionality for the main form. I have created a
method that checks for valid input among input fields and have created
a method for each input field that is run when the user leaves the
input field. These methods then call the check input method to valid
user input.