ContactList3 Commit 8
This Commit provides additional functionality for the main form. I have
created 2 methods for interacting with the contact list information.
Once a user has double clicked a list view item the apply changes and
delete buttons are made visible. There is a method to delete a contact
from the list. There is a method to apply the changes made to a contact
in the input fields.