TicTacToe3 Commit 1
This commit provides initial functionality for the game. There are
multiple member variables created to hold data such as image index,
iteration of the game, what game pieces have been played, and which
player played which spot. There are 2 methods that set the image index
based on whether player 1 is X or O.