TicTacToe3 Commit 4
This Commit provides additional functionality for the game. There are 2
new methods, a new game method that resets all of the game data and the
game board, and a new game button click method. This method simply
calls the new game method. There are 2 separate methods because there
is more than one instance where a new game must be created.