TicTacToe3 Commit 5
This Commit provides the main functionality for the game. There are 9
new methods, one for each game piece button click. When a game piece
button is clicked the method first checks that it hasn’t been played
yet, then it calculates which player clicked it based on the iteration
and player image index. Each method then stores valuable game data.