TicTacToe3 Commit 6
This Commit provides additional functionality for the game. There are 2
new methods. There is a method to check the score and one to display
whether or not a player has won. The check score method goes through
the 2d array of game data that store which player played where. If
there are 3 in a row then the show winner method is called to displayed
who won. If the iteration reaches its maximum and no player has won
then show winner displays that the game was a draw.