TicTacToe3 Commit 8
This Commit provides additional functionality to complete the game.
There are 2 new methods one for the load button click and one to
configure the loaded XML data. The load button click method uses an XML
reader to create a path and iterates through the XML nodes to retrieve
game data. At the end of this method the configure XML data method is
called and given the two game data lists in string format. The
configure XML data method takes the loaded data and sets the game data
to the loaded data then updates the game board to the loaded game.